+Add post_product_[uv] and post_product_sum_[uv]
  Added four new routines to within MOM_diag_manager.F90 to write out a
diagnostic based on the product of two arrays at velocity points, either as a
3-d diagnostic or a 2-diagnostic of the vertical sum.  Doing it this way both
simplifies the higher level code and promotes the reuse of scratch arrays for
diagnostics.  In tests that use these new interfaces, diagnostics are identical
to their previous values.  All answers are bitwise identical, but there are 4
new public interfaces.