+Rename G%Rad_Earth in grid types
  Renamed G%Rad_Earth to G%Rad_Earth_mks in both the ocean_grid_type and the
dyn_horgrid_type to deliberately cause any code using these variables instead of
G%Rad_Earth_L to fail to compile as a step toward eliminating the un-scaled
variable.  Also modified the two places where the mct and NUOPC drivers were
using G%Rad_Earth to instead use G%Rad_Earth_L.  Were it not for the fact that
there could be code that is using this variable that is not visible to the main
branch of MOM6, this variable would have been simply eliminated.  All answers
are bitwise identical, but there is a deliberate change to an element in two
important transparent types.