Use G%Rad_Earth_L in mct and NUOPC cap code
  Modified the two lines where the mct and NUOPC caps were using G%Rad_Earth to
instead use G%Rad_Earth_L, and cancelled out the dimensional rescaling factors
on these lines that are no longer appropriate.  All answers should be bitwise
identical, but because this change is to code that is not routinely exercised by
testing with MOM6-examples, extra care should be taken in assessing these
changes.