+Add US args and rescale dt arg to generic tracers
  Added unit_scaling_type arguments to various routines that had previously used
a unit scaling type, but did so via the G%US pointer, to make the type
dependencies more explicit and to avoid unnecessary pointer use.  It had been
the intention to make these arguments explicit from the time they were
introduced via a pointer in the ocean_grid_type as a temporary convenience.
The construct G%US%... was replaced with US%... wherever it was possible.

  Also rescaled some local variables or corrected comments in oil_tracer.F90,
nw2_tracers.F90, and boundary_impulse_tracer.F90, and rescaled the units of
the dt argument to MOM_generic_tracer_column_physics from [s] to [T ~> s].

  All answers are bitwise identical, although there are multiple changes to
public interfaces.