Use global_mass_integral in lateral_bdry_diff
  Use global_mass_integral for the debugging diagnostics of the tracer amounts
before and after diffusion in lateral_boundary_diffusion, and replaced a call to
write(*,*) with a call to MOM_mesg to actually write the message.  The
global_mass_integral uses reproducing sums, and is invariant to layout, while
MOM_mesg is preferable for output because it will allow us to more cleanly
control how output is handled and which processors do the writing.  All
solutions are bitwise identical, although some debugging output will change.