Corrected the license on RGC_initialization.F90
  Corrected the license on RGC_initialization.F90 to refer to the same
LICENSE.md file as the rest of the MOM6 code, with the explicit written consent
of the contributors to this file.  Also changed the local variable "mod" into
"mdl" in this module to avoid some of the problems we had previously found with
some compilers from reusing an intrinsic function name for a variable name.  All
answers are bitwise identical.