Add a deta_dt diagnostic to the split BT/BC mode
The SSH tendency can now be computed for the split baroclinic/
bartropic (integrated with RK2) way of calculating the dynamics.
This extra diagnostic is needed for computing vertical mass
transport (wmo) in z* coordinates offline from the convergence
of the horizontal mass transports.

The meaning, units, and conversion factors associated with the
deta_dt diagnostic change between Boussinesq and non-Boussinesq
modes. This commit accounts for these differences and reports the
units consistent with the mass transports.