+Convective adjustment outside regridding_main
  Eliminated the option to do convective adjustment within regridding_main,
including making the optional argument conv_adjust mandatory, with an error
message if it is set to .true., so that there will not be the possibility of
this code change silently changing the code behavior or solutions.  After a
decent interval, this argument can be safely eliminated.  As a result of these
changes, the intent of two key arguments to regridding_main() could be changed
from intent(inout) to intent(in), which should help clarify the purpose of this
routine. All answers are bitwise identical, but there are interface changes.