Fix non-standard white space (#109)
* Fix non-standard white space

  Made widespread corrections to indentation and other white space issues to
conform to the MOM6 standards, as documented in the white space section of the
MOM6 style guide, at https://github.com/mom-ocean/MOM6/wiki/Code-style-guide.

- MOM6 code uses a 2-point indent scheme.

- There should be a space around all assignment (" = ") operators.

- The names of optional arguments do not use whitespace, to differentiate them
  from assignments.

- There is a space around the semicolons for stacked enddo statements.

After this change, greping for '^ [a-zA-Z]', '^   [a-zA-Z]' or '^     [a-zA-Z]'
does not find any lines in the MOM6 src or config_src/infra code.  All answers
and output are bitwise identical.