Impose standard indentation in most driver code
  Corrected unbalanced indentation levels and impose MOM6-standard 2-point
indentation in the config_src/driver codes, including both the FMS and mct caps
and the solo_driver.  The nuopc_cap is excluded from this commit because of the
huge number (over 1000) of lines there that use irregular indentation.  All
answers are bitwise identical, and only white space changes are included.