(*)Internal thickness variable unit correction
  Changed the internal units of several internal thickness variables in
ePBL_column and of some temporary diagnostics in MOM_bulk_mixed_layer from [Z]
to [H] to reduce the number of rescaling factors that are being applied outside
of the conversion factors when diagnostics are registered.  All answers and
output are bitwise identical in the MOM6-examples test suite, although in some
cases where ePBL is used with a non-Boussinesq model, there may be answers at
roundoff.  I wanted to make this change now before any such models exist.