Simplify set_grid_metrics_from_mosaic
  Simplified the code in set_grid_metrics_from_mosaic and set_coord_from_file
using the scale argument to MOM_read_data, eliminating several arrays that are
no longer needed.  Also added a call to set RAD_EARTH via get_param for grids
where it is not otherwise set, replacing a hard-coded constant.  The variable
G%Rad_Earth is used in a few places even in these cases, so it should be logged
as a run-time parameter.  All answers are bitwise identical, although there are
new entries in some MOM_parameter_doc.all files.