(*)Fix occasional bug with PFv_visc_rem diagnostic
  Corrected a bug that causes the bottom-drag filtered meridional acceleration
diagnostics, like PFv_visc_rem, to erroneously be filled with all zeros unless
one of the filtered u-point diagnostics is also enabled in the diag_table, and
conversely could lead to segmentation fault if there are are only filtered
u-point diagnostics.  However, if at least one of both kinds of diagnostics are
enabled, everything was already working as intended.  Also changed a call to
enabled_averaging into a call to enable_averages to avoid having to scale an
argument.  All solutions are bitwise identical, and in many cases the
diagnostics are also unchanged.