Rescale ice shelf temperatures
  Applied dimensional rescaling of all of the ice shelf temperature and salinity
variables that could be modified without requiring any changes at the driver
level.  There is a new unit_scale_type argument to MOM_IS_diag_mediator_init,
mirroring what is done in MOM_diag_mediator_init.  This commit also includes
changing the initial values in the stencil array from -1 to 0, but this
initialization appears to be unnecessary.  A large number of comments were
added, corrected, or modified.  All answers in the test cases are bitwise
identical.