CI: Restructure of GitHub Actions
This patch separates actions/testing-setup/action.yml into separate
files for general behavior and platform-dependent installations.

All workflows excepting the MacOS test use the Ubuntu setup.

The Mac OS test was also changed from the stencil tests to a regression
test.

This could be worked on more to set it up like a matrix, and to run
across multiple platforms, but this is good enough for small extensions
to platforms like Mac OS.