(*)Avoid array syntax math in MOM_wave_structure
  Revised the MOM_wave_structure code to avoid using array syntax in
calculations and subroutine arguments.  While making these changes, several
incorrect or missing variable unit descriptions in comments were identified and
corrected.  This set of changes is mathematically equivalent, but will result in
changes at the level of roundoff, either because the order of sums was changed
or because divisions are replaced by multiplication by a common reciprocal.
Some debugging safety checks were moved inside of a logical test of the debug
flag for efficiency.  One dimensionally inconsistent expression (in vertical and
horizontal distances) was corrected, so that it should now pass the dimensional
consistency checks.

  This commit will change answers in any cases that depend on this code, but
because we recently corrected (in MOM6 PR#154) a major bug in this code with a
much larger impact without causing any disruptions, we can be confident that
this code is not yet used in an production runs where minor answer changes could
be problematic.  This code change does not alter any answers or output in the
MOM6-examples test suite.