+Restored public interface to post_data_1d_k
  Made post_data_1d_k publicly visible once again, rather than requiring calls
to use post_data, to support backward compatibility with older versions of the
ocean_BGC code.  This interface was removed from public visibility as a part of
github.com/NOAA-GFDL/MOM6/pull/107, but it caused problems with some of GFDL's
Earth System Models, as noted in https://github.com/NOAA-GFDL/MOM6/issues/168.
All answers are bitwise identical for any cases that compiled before.