Implement spherical harmonics SAL from MPAS-Ocean
This is an initial step to add the functionality of calculating
self-attraction and loading (SAL) with the spherical harmonic transform
(SHT) method. The code in this commit is a direction translation and
adoption of the "parallel SAL" from MPAS-Ocean model developed at the
Department of Energy's Los Alamos National Laboratory
(Barton et al. (2022) and Brus et al. (2022)).

This commit serves as an "MVP" so that the MPAS-Ocean algorithm is
functional in MOM6. Further refactor, optimization and documentation
should be put in place.

* A new module MOM_spherical_harmonics is introduced to house
precomputed coefficients used for SHT.
* New subroutines are added in module MOM_tidal_forcing for calculating
SAL with the SHT method.