+Use answer_date to specify remapping
  Replace answers_2018 arguments with answer_date arguments to specify the
version of expressions used in a number of vertical or horizontal regridding
calls.  In some cases, this also involves replacing one of the elements in an
opaque type.  It can also involve reading (but not yet logging) the new runtime
parameter DEFAULT_ANSWER_DATE, but if it not set the results are unchanged from
before.  There is also a new optional argument, remap_answer_date, to
wave_speed_init and wave_speed_set_param.  Some comments were also added to
describe real variables in VarMix_init.  All answers are bitwise identical.