(*)Set defaults in USE_REGRIDDING get_param calls
  Set default values in all get_param calls for USE_REGRIDDING.  Previously,
there had been 4 calls where this was missing, which led to the problems noted
at https://github.com/mom-ocean/MOM6/issues/1576.  This PR will allow that issue
to be closed.  Also used the default argument in a get_param call for INPUTDIR,
although that case would not change any behavior because the value was set
before the get_param call.  A fail_in_missing argument was added to the FMS_cap
call to get_param for GUST_2D_FILE, mirroring what is done for the solo_driver
code, but cases where this was actually missing were very likely to have failed
later anyway, but without an explicit error message.  This PR could change
unpredictable behavior in cases where USE_REGRIDDING is not explicitly set, but
all answers are bitwise identical in the MOM6-examples test suite.