Refactor database_client_interface stub
The database client stub used Fortran 2018 features which do not
adhere to the MOM6 allowing only up to Fortran 2008. To fix this,
separate interfaces for tensors with 1-4 dimensions replace the
assumed rank subroutines. This does not interfere with the
SmartRedis library as the public (overloaded) interface still
retains the same API.

Additionally, some methods which are likely unique to the
SmartRedis client have been removed to enhance the likelihood of
providing a general database client API to be used by other
implementations.