Point gitmodules to mom-ocean forks
This patch modifies the URLs of the git modulues from the main
repositories to local mirrors at @mom-ocean.

Since the MOM6 sourc code is hard-coded to particular hashes, there is
no benefit to linking to the most recent versions of these codes.  And
linking to local forks controlled by us will protect us from potential
removal of the commits in these repositories.