+Add only_read_from_restarts
  Added the new overloaded interface only_read_from_restarts to allow for
variables that are not registered to be stored in restart files to be read from
the restart files.  Versions of this routine that work on 2d, 3d or 4d arrays
and pairs of 3d arrays have been added for now, although the pattern should be
clear enough if other sizes of arrays are needed.  They use the new private
routine find_var_in_restart_files to determine whether a named variable is in
the restart files and the full path to the appropriate file. These routines
allow for the contents of the restart files to be gracefully altered, while
preserving the ability to use restart files with the previous format.  All
answers are bitwise identical, but there is a new public interface.