+Add new runtime parameter STORE_CORIOLIS_ACCEL
  Added the option to store the Coriolis and advective accelerations from the
end of one split RK2 dynamics time step for use in the predictor phase of the
next time step, in order to allow for a future commit that will properly remap
these accelerations in ALE mode.  This change is controlled by the new runtime
parameter STORE_CORIOLIS_ACCEL, and it leads to changes in the fields that are
saved in the restart files.   This commit also includes the necessary changes to
allow the restart files written in one mode to be read in the other and give
bitwise identical restarts.  Because the next step's predictor CAu and CAv
accelerations now have to coexist with the previous step's CAu and CAv
accelerations, which are used for derived diagnostics (like energy budget
terms), the predictor CAu and CAv terms have to be stored in new variables,
CS%CAu_pred and CS%CAv_pred, and there is a new accel_diag_ptrs type in the
dyn_split_RK2 control structure to diagnose truncation errors arising in the
predictor step.  Several comments have also been modified to document the
dimensions of variables.  All answers and diagnostics are bitwise identical.