State-dependent options for diag coordinates
Reference pressure and the compressibility fraction could
previously only be set for coordinates if they were used as the
main prognostic coordinate. This updates the logic to allow these
parameters to be changed for diagnostic coordinates. To avoid,
naming clashes with the prognostic coordinate, new function was
written to construct the correct parameter name. Other places
within the codebase where this occurred have subsequently been
refactored.