(*)Update remap_answer_date set in uninitialized types
  Updated the default values of remap_answer_date as declared during the
specification of the wave_speed_CS, remapping_CS and regridding_CS to 99991231
(to use the very latest version of the algorithms) if these control structures
are used without further initialization via optional arguments to their various
initialization routines (wave_speed_init(), initialize_remapping() and
initialize_regridding() or set_regrid_params()).  In testing both via the TC
tests and the MOM6-examples regression suite, all answers are bitwise identical,
because every instance where these are used either does have an appropriate call
to the initialization routine, or because (as is the case for remapping
diagnostics) they are hard-coded to use the PLM remapping scheme, which is not
impacted by this answer_date.  It is conceivable, if unlikely, however, that
there could be cases outside of the standard MOM6 code that are using this
remapping capability without proper initialization of this flag, in which case
answers could change.  In addition, a comment was updated in the interp_CS_type
indicating the cases that would be altered by changing the hard coded default
answer_date there, but the default in that case was not changed.  All answers
and output are identical in the MOM6-examples and TC test suites.