+(*)Fix rescaling issue with MOM_stoch_eos
  Added a missing dimensional rescaling factor in the get_param calls for
KD_SMOOTH in MOM_stoch_eos_init and another for the recently added variable
DT_OBC_SEG_UPDATE_OBGC in initialize_MOM, both of which would have caused
certain cases to fail the dimensional consistency tests.  This correction
required the addition of a new unit_scale_type argument to MOM_stoch_eos_init,
and this change in interface presented an opportunity to reorder the arguments
to this routine to match the order used in every other init call.

  There was also some modest refactoring of the MOM_stoch_eos code.  Four arrays
in the MOM_stoch_eos_CS type that had been declared to optionally use static
memory allocation were modified to be simple allocatables, as they are not used
in the vast majority of MOM6 cases, and there is no reason to always assign
memory to them.  Also, the register_restart_field call for "stoch_eos_pattern"
was revised to use the newer, more direct form rather than working via a vardesc
type.  The comments describing several real variables in this module were added
or modified to describe their units using the standard format.  The module use
statements at the start of this module were updated to reflect these changes.

  All answers are bitwise identical in cases that do not use dimensional
rescaling, but answers will change (be corrected) in some cases that do use
dimensional consistency tests.  The public interface to MOM_stoch_eos_init was
altered.