Use small minimum ustar in mixed layer restratification
If the forcing ustar is exactly zero, the denominator of the momentum
mixing rate term reduces to just the Coriolis parameter. With a grid
construction that is symmetric about the equator, this term is also
exactly zero, leading to a division by zero.

To avoid this, a very small minimum ustar value is introduced, using
the Earth's rotation as a velocity timescale, in the same manner as in
some of the vertical parameterisations. This should prevent the
denominator of the momentum mixing rate from going to zero.