.testing: Fix concurrency errors in tc4 rules
This patch fixes issues in the preprocessing of tc4.  Specifically, we
were directing one rule to produce two output files, which resulted in
the rule being run twice when invoked in parallel (make -j).

This has been replaced with the recommended solution for handling
concurrent outputs: use one to generate both, and connect the second to
the first with a separate rule.

I also generalized the `make` command in the .testing Makefile.

This should address (and hopefully fix) some intermittent errors in the
.testing build on Gaea.