+Add x_ax_unit_short to ocean_grid_type
  Added the new elements x_ax_unit_short and y_ax_unit_short to the
ocean_grid_type and the dyn_horgrid_type to facilitate the documentation of the
units of variables when they depend on the units of the axes.  This includes
modifying some units that were being reported as the meaningless "[k]" into the
meaningful "[km]", and some units that were being reported as just "[degrees]"
are now being reported with the more specific "[degrees_E]" or "[degrees_N]".
These new elements are also being used in the baroclinic_zone_initialization and
circle_OBCs modules, and comments describing the units of a number of the
internal variables in the former module were added or modified.  The description
of the AXIS_UNITS parameter was expanded to provide some of the other working
options.  All answers are bitwise identical, but there are minor changes in a
number of the MOM_parameter_doc files.