(*)Correct USE_STANLEY domain extents in EOS calls
  Corrected the domain extents used in EOS calls that are triggered with
USE_STANLEY_GM and USE_STANLEY_ISO.  These bugs were accidentally introduced
when the changes adding the MOM_stoch_eos code to the main branch of MOM6 were
merged with changes on the dev/gfdl branch.  Also added a test for cases when
USE_STANLEY_GM is set to true but STANLEY_COEF is negative to reset the internal
versions of this flag to false with a sensible warning message rather than
encountering segmentation faults.  All solutions are bitwise identical in cases
that worked before.