Add GLwork diagnostic
New diagnostics:
* GLwork: Kinetic Energy Source from GL90 Vertical Viscosity

The energetics of the GL90 parameterization (named "GLwork") are intentionally computed in
MOM_vert_friction, rather than in MOM_diagnostics, where the reamining kinetic energy budget
terms are computed. We have to do the computation in MOM_vert_friction to ensure sign-
definiteness when GLwork is summed in the vertical. Indeed, MOM_diagnostics does not have
access to the velocities and thicknesses used in the vertical solver, but rather uses a time-mean
barotropic transport [uv]h to compute the energy budget diagnostics. A detailed discussion and
exploration of this issue can be found in https://github.com/ocean-eddy-cpt/MOM6/issues/25.

As a result of not computing the energetics in MOM_diagnostics, GLwork is not exactly contained
in KE_visc. KE_visc represents the energetics of all vertical viscosity contributions, including
the GL90 vertical viscosity. We could implement a term "KE_visc_gl90" that can be 1-to-1 compared
to KE_visc; that is, KE_visc - KE_visc_gl90 would represent exactly the energetics of all viscosity
contributions EXCEPT the GL90 viscosity. If we implemented KE_visc_gl90, this term would in practice
be very similar as GLwork, but sign-definiteness is not ensured, see above.