Introduce logical variable KD_GL90_USE_EBT_STRUCT
This variable is analogous to KHTH_USE_EBT_STRUCT, but is specifically for the GL90 scheme.
If the user sets KD_GL90_USE_EBT_STRUCT = True, an EBT structure will be applied to KD_GL90.
Before this commit, KHTH_USE_EBT_STRUCT took the two-fold role of imposing an EBT structure
on KHTH (used in the GM scheme) and KD_GL90 (used in the GL scheme), which may have been confusing.