+Pass ocean_grid_type to call_tracer_register
  Pass ocean_grid_type arguments to call_tracer_register and call_OBC_register
in place of the hor_index_type arguments that had been used previously.  Also
use these ocean_grid_type arguments in calls to register_shelfwave_OBC,
register_RGC_tracer and register_advection_test_tracer.  Within these three
routines, the contents of the ocean_grid_type are used to specify axis units.
The new runtime parameter SHELFWAVE_AMPLITUDE was added to allow for run-time
control of the amplitude of the shelfwave test case.  By default all answers are
bitwise identical, but there are some changes in the units of parameters as
documented in the MOM_parameter_doc files and a new entry in these files for the
shelfwave test case.