+Revise interfaces to horiz_interp_and_extract
  Revised the interface to the two horiz_interp_and_extract to eliminate the
unused reentrant_x and tripolar_n arguments and to rename the conversion
argument to scale and make it the last mandatory argument in anticipation that
we might decide to make it optional, similarly to other routines like
MOM_read_data.  The renaming is because we use 'conversion' to indicate how the
internal representation is to be rescaled for output, most prominently in
register_diag_field, whereas everywhere else where we are rescaling input to the
model's internal units, we use a scale argument.  This makes the convention
self-consistent across the MOM6 code, and should avoid some confusion.  The
calls to these routines were updated in 8 places in 4 modules.  In 8 places the
get_param calls for TRIPOLAR_N or REENTRANT_X that are no longer needed were
eliminated as were the associated internal variables.  This commit also includes
some additions to comments near the changes that were directly tied to this
commit.  All answers are bitwise identical, but there are changes to a publicly
visible interface; code that tries to use the old interface will not compile.