makedep: Include object file dependencies
The current implementation of makedep contains something like a race
condition, where the creation of the .mod and .o files may be in an
order which breaks the current dependency tree.  Currently, .mod depends
on .o, and changes to a module do not trigger rebuilds of dependent
source.

Rather than try to sort out the rule order, which could even depend on
compiler internals, this patch just adds both object and module output
files as dependencies, and rebuilds if either changes.

We might want to come back to this someday and understand the actual
order of rule execution.

Thanks to Alistair Adcroft (@adcroft) for proposing this solution.