*+Use runtime params in adjustment_initialization
  This commit replaces previously hard-coded values for dRho_dS and the
temperature range for the sloshing test case with values read in from the
existing runtime parameters DRHO_DS and the previously unused parameter T_RANGE
that was already being read in for this configuration.  The default value for
dRho_dS from the EOS code (0.8 kg m-3 ppt-1) does not match hard coded value
here (1 kg m-3 ppt-1), but in the existing tests using adjustment_initialization
DRHO_DS was being explicitly set to the hard-coded value here, so I have chosen
to reuse the same parameter with the same default value as is used elsewhere
even though it could change answers in cases that do not set DRHO_DS
consistently with the previous hard coded parameter.  In addition, the previous
default value for T_RANGE read into this module (0 degC) did not match the
hard-coded range that was being used (1 degC).  In this case I have chosen to
change the default value for T_RANGE in this module so that by default this will
reproduce the previous answers.  As a result, there will be changes in the
MOM_parameter doc all files for cases (like adjustment2d) that use the
adjustment_initialization routines.  This commit will change answers in some
cases, but in the existing adjustment2d test cases in the MOM6-examples test
suite, the answers are bitwise identical without any changes to the existing
input parameter files.