(*)Require dz_subML to calculate the sub-ML N2
  Introduced a fatal error when diagnoseMLDbyDensityDifference is directed to
calculate the mean stratification in a region below the mixed layer but dz_subML
is not provided as the depth extent over which to average the stratification,
replacing the previous hard coded (but unused) default value of 50 m.   Also
replaced a hard coded floor on the minimum thickness over which to apply
boundary fluxes of 1e-30 m in applyBoundaryFluxesInOut with GV%H_subroundoff,
which is also a tiny value.   This floor is here just to avoid division by 0;
while it is possible for this to change answers in some cases where ANGSTROM=0,
even then there are other limiters that should apply so the solutions should
not change.  All answers are bitwise identical with for the MOM6-examples test
suite, and very likely for all existing MOM6 configurations.