+Make z_tol arg non-optional to cut_off_column_top
  Made the previously optional z_tol argument to find_depth_of_pressure_in_cell
and cut_off_column_top non-optional.  It was already being provided except in a
call for unit testing, so adding it as a parameter there led to a simple change
and the elimination of a hard-coded dimensional parameter.  Also replaced the
hard-coded fill values over land in MOM_temp_salt_initialize_from_Z for
temperatures and salinities before regridding with the runtime variables
temp_land_fill and salt_land_fill that are already being used in the same
routine.  This does not change any answers, probably because these values are
not actually used.  All answers are bitwise identical, but some subroutine
arguments have been made non-optional.