+Obsoleted the runtime parameter NEW_SPONGES
  Formally obsoleted the runtime parameter NEW_SPONGES.  The agreed upon
replacement is INTERPOLATE_SPONGE_TIME_SPACE, which has been available for
almost a year.  There is a warning message rather than a fatal error if
NEW_SPONGES is used and both are set consistently, and a hint if they are not.
Also added or amended comments describing a number of the internal variables or
their units in MOM_state_initialization.  All answers and output are bitwise
identical.