+(*)Fix USE_QG_LEITH_VISC = True reproducibility
  This commit includes a series of changes that get MOM6 to reproduce across
processor counts, layouts and restarts when USE_QG_LEITH_VISC = True and also to
implement related changes to the QG Leith code to be consistent with not making
the Boussinesq approximation when in non-Boussinesq mode.  These changes
include:

  - Adding a thermo_var_ptrs and a timestep argument to to the interfaces to
    horizontal_viscosity, initialize_dyn_split_RK2 and initialize_dyn_split_RK2b

  - Correcting a bug in the powers of H_subroundoff in the denominators of some
    expressions for the effective thicknesses associated with interfaces,
    thereby correcting some dimensionally inconsistencies in these expressions

  - Adding the new publicly visible routine calc_QG_slopes

  - Adding the arguments slope_x and slope_y to calc_QG_Leith_viscosity and
    avoiding the use of the VarMix%slope_[xy] elements in that routine

  - Using thickness_to_dz to consistently convert layer thicknesses to the
    vertical distances across layers and provide this as an argument to
    calc_QG_Leith_viscosity

  - Determining vertical distances in calculating the vertical derivatives of
    the slopes consistently with avoiding the Boussinesq approximation when in
    non-Boussinesq mode

  - Increasing some loop extents in calc_QG_Leith_viscosity

  Note that several of the expansions of the halo sizes in updates or in some of
the calculations are not necessary if an extra halo update is included for
slope_x and slope_y and dz after the calls to calc_QG_slopes and thickness_to_dz
in horizontal_viscosity.  This extra halo update has also been included in a
comment for later reference.

  This commit includes several changes to publicly visible interfaces, and it
will change answers with USE_QG_LEITH_VISC = True (which had previously
triggered a fatal error), but answers and output are bitwise identical in other
cases.