Use unscale args to check_redundant calls
  Added unscale arguments to the various check_redundant calls so that any error
messages that are generated by inconsistent redundant points will be invariant
to the unit scaling that is is in use.  Also rescaled the units of
dtbt_reset_period to [T ~> s] in the MOM_control_struct.  All answers are
bitwise identical, but in some rare debugging output will become consistent.