+Update check_MOM6_scaling_factors with C and S
  Updated check_MOM6_scaling_factors and compose_dimension_list to reflect that
fact that MOM6 is now doing dimensional consistency testing for temperature (via
[C ~> degC]) and salinity (via [S ~> ppt]), with an expanded dimension of the
scaling key from 6 to 8 and additional calls to add_scaling.  Also updated the
weights on the add_scaling calls, which are essentially counts of the frequency
of the various unit descriptors in the MOM6 code, to reflect only the counts of
variables with doxygen comments (i.e., arguments, function return values and
elements of types) but excluding the user, framework and diagnostics directories
and the passive tracer packages.  All model solutions are bitwise identical, but
there will be updated suggestions for combinations of scaling factors that
minimize the aliasing of the units that are used.