+(*)Remove coord_SLight
  Removed the coord_slight module and all calls to it, and obsoleted all
run-time parameters that are exclusively related to it.  This code was an
attempt from 2015 to define an appropriate hybrid vertical coordinate for global
climate modeling, but it never worked very well (usually falling apart in the
second year), and it has not been used in any publication or active model for
many years.  The test case that exercised this coordinate in the MOM6-examples
test suite is also being removed via MOM6-examples PR #388.  The coord_SLight
code is being eliminated altogether now to simplify the MOM6 code base and
reduce the volume of untested and unused code.  All answers in all known MOM6
configurations in active use are bitwise identical, although there is a remote
chance that someone somewhere might be using the SLIGHT coordinate.