+Make values arg to get_field_nc intent inout
  Made the values argument to MOM_read_data_2d and get_field_nc intent inout to
preserve the previous values of this array in any halos it might have, and
removed a line initializing the halos of this array to 0.  Also added some
missing comments describing arguments to these routines  and changed the
existing comments describing the arguments to get_field_nc to use doxygen
syntax.  All answers in test cases are bitwise identical.