Renaming input parameter names for the SAL module
Remove "TIDAL" and "TIDE" from the relevant input names of the SAL
module, as SAL is not specific to the tides. This affects both scalar
approximation and the fully online spherical harmonic options.

* For scalar SAL, old parameter names are still acceptable, but a
WARNING is given if these old names appear in MOM_input.
* For read-in SAL, no change is made.
* For iterative method (use_prev_tides), the use of
TIDE_SAL_SCALAR_VALUE is completely deprecated, as this is a feature
that is rarely used.
* For harmonic SAL, a relatively recent feature, a hard obsolete is
applied, i.e. if the old parameters are specified, a FATAL error is
given, unless the new parameters also exist and match the values of the
old parameters.

List of input names changed:
* TIDE_USE_SAL_SCALAR -> USE_SAL_SCALAR
* TIDE_SAL_SCALAR_VALUE -> SAL_SCALAR_VALUE
* TIDAL_SAL_SHT -> USE_SAL_HARMONICS
* TIDAL_SAL_SHT_DEGREE -> SAL_HARMONICS_DEGREE