*Use tv%SpV_avg in non-Boussinesq regridding
  Use SpV_avg in both hybgen_unmix and regridding_main to estimate the nominal
ocean bottom depth in thickness units when in fully non-Boussinesq mode.  This
change eliminates certain dependencies on the Boussinesq reference density via
GV%Z_to_H.  Also added a call to calc_derived_thermo in ALE_regrid_accelerated
that is necessary for the specific volume to be updated, along with tests for
the validity of the specific volume with the 1-point halos that are currently
used with ALE regridding and remapping.

  Also, use the total thickness place of the nominal depth in build_grid_rho and
build_grid_sigma when in fully non-Boussinesq mode.  This is mathematically
equivalent but changes answers at roundoff.

  All answers are bitwise identical in Boussinesq mode, but ALE-based solutions
change in fully non-Boussinesq mode with this commit.