Resolve warning about S_REF units
The units of S_REF were first registered in "PSU" and then later by
benchmark in "ppt" which led to a WARNING. Although there is no uniformity
to the units of S_REF throughout the user code (currently PSU, ppt, and 1e-3)
are all used, benchmark is the only one that leads to a warning in the current
suite and changing all to be consistent would 1) unnecessarily update the doc
files, and 2) not be correct for all models. Bottom line, I'm punting on
the best way to handle units of salinity. This commit resolves the only WARNING
we currently get in the MOM6-examples suite.