+Rename OBC sea surface height variables
  Renamed OBC related variables to emphasize that they are sea surface heights,
and not sea surface heights or total column mass depending on the Boussinesq
approximation.  Also changed the units of these renamed variables to [Z ~> m]
instead of [H ~> m or kg m-2].

  The renamed element of the OBC_segment_type is eta (which is now SSH).  The
renamed and rescaled elements of the BT_OBC_type are H_[uv] (which are now
dZ_[uv]) and  eta_outer_[uv] (which are now SSH_outer_[uv]).  The internal
variables H_[uv] and h_in in apply_velocity_OBCs are now dZ_[uv] and ssh_in.
Tidal_elev in update_OBC_segment_data and cff_eta in tidal_bay_set_OBC_data
were rescaled but not renamed.  There is also a new vertical grid type argument
to apply_velocity_OBCs for use in changing the scaling of the SSH variables.

  A total of 11 GV%Z_to_H or GV%H_to_Z rescaling factors were cancelled out as a
result of these changes, while 16 new ones were added, but most of these will in
turn be dealt with in a follow-on commit that enables the use of OBCs in
non-Boussinesq mode.

  Because any cases that use Flather open boundary conditions with the
non-Boussinesq mode issue a fatal error, all answers are bitwise identical, but
there are changes to the names and units of elements in a transparent type.