*Fix logic of an inconsistent initialization test
  Corrected the logic of a recently added test for inconsistently specified
thickness units during the initialization of a new run to only apply to a new
run.  This was causing an incorrect fatal error with some restart tests.  All
answers are bitwise identical in cases that worked previously, but it corrects
the problem with the restarted cases that had been aborted by the test that
was added with the revision to the Rossby_front_2d_initialization code.