*Non-Boussinesq revision of kappa_shear
  Remove dependency on the Boussinesq reference density in MOM_kappa_shear when
in fully non-Boussinesq mode.  This is done by using calls to calculate_density
and calculate_specific_vol_derivs to calculate the thermal expansion and haline
contraction coefficients, and by using thickness_to_dz to convert layer
thicknesses to vertical distances.

  A large part of the changes to the kappa_shear code involved refactoring it to
differentiate between thicknesses and geometric heights by working internally
with dynamic viscosities and diffusivities in non-Boussinesq mode, and to keep
thicknesses in thickness units rather than converting them to distances in
height units.  The internal variable dz was renamed to h_lay in
Calculate_kappa_shear and Calc_kappa_shear_vertex for greater clarity.

  The scaling factor for dz_massless is now set to (US%Z_to_m*GV%m_to_H) in
Calculate_kappa_shear to reduce the dependency on the Boussinesq reference
density; this does not impact answers if RHO_0 = RHO_KV_CONVERT.

  The internal changes related to this are extensive.  This commit includes
changing the units of 63 internal variables, and the addition or renaming of 17
internal variables.  There are 2 new arguments to find_kappa_tke (one is
necessary to differentiate thicknesses from vertical distances and while the
other is used to preserve Boussinesq-mode answers), and 1 new argument to
kappa_shear_column plus the renaming of another.  The units of 13 internal
subroutine arguments and 3 elements of the Kappa_shear_CS type change. A total
of 6 thickness unit conversion factors were eliminated with these changes.

  The answers will change in non-Boussinesq mode when the Jackson et al. shear
mixing parameterization is in use, but they are bitwise identical in all
Boussinesq test cases.  There are no changes to external interfaces.