*Non-Boussinesq revision of 3 MOM_CVMix modules
  This commit revises MOM_CVMix_conv, MOM_CVMix_ddiff and MOM_CVMix_shear to
work in an appropriate mixture of thickness and vertical extent variables to
enable their use in non-Boussinesq mode, using thickness_to_dz to convert
between the two, while retaining the previous answers in Boussinesq mode.  This
includes the use of a layer thicknesses rather than a vertical distance in the
denominator of the calculation of the buoyancy frequency and to the units of the
internal rescaled gravity variable in CVMix_conv.  This commit eliminates any
direct or indirect dependency on the Boussinesq reference density in these 3
modules when in non-Boussinesq mode.

  This set of changes includes changing the units of 7 internal variables and
the addition, renaming, or change in the dimensions of 4 internal variables.
One element in the CVMix_ddiff_cs type is rescaled to use thickness units.
These changes lead to the removal of 4 factors of GV%H_to_Z in these 3 modules.

  Answers will change in non-Boussinesq mode when USE_CVMix_CONVECTION,
USE_CVMIX_DDIFF, USE_PP81 or USE_LMD94 are true, but they are bitwise identical
in all Boussinesq test cases.