*Fix allocate_forcing_by_ref tau_mag_gustless bug
  Added the name to the do_taumag argument in a call to allocate_forcing_type
in allocate_forcing_by_ref to account for the fact that there are unused
wave-related optional arguments in this interface.  When this was omitted in
the current code, the wrong arrays are being allocated during rotation tests
with resultant segmentation faults in those tests.  This commit corrects a bug
that was recently added with MOM6 dev/gfdl PR #445.