+Remove rescale_grid_bathymetry
  Removed the unused routine rescale_grid_bathymetry.  This routine was added in
August 2018 as a part of the development of the depth unit conversion and
dimensional consistency testing, but it is no longer being called now that this
conversion is essentially complete (and it has not been called by the code in
several years).  For the original commit that first added this code, see
github.com/mom-ocean/MOM6/commit/ddc9ed1c33a1b7357b213929118ecaa19ae63f9f.  All
answers are bitwise identical.