*+Use TIDES_ANSWER_DATE with semi-Boussinesq tides
  Use TIDES_ANSWER_DATE with an additional if-block to recover the dev/gfdl
answers in semi-Boussinesq mode runs with tides enabled, similarly to what is
already being done in Boussinesq mode.  This changes some answers for the
nonBous_global and tides_025 test cases back to those that would be obtained
with the code on dev/gfdl.   TIDES_ANSWER_DATE is now logged in all cases with
tides, so the MOM_parameter_doc files change in some non-Boussinesq cases.