+*Add halo_size argument to wave_speeds
  Replace the optional full_halos logical argument to wave_speed and wave_speeds
with an optional halo_size integer argument, following the pattern used
elsewhere in the MOM6 code, with a similar change to itidal_lowmode_loss.  This
halo_size argument is used in the call to thickness_to_dz in wave_speeds, and
the call to wave_speeds in propagate_int_tides was modified accordingly.  In
addition, the loop bounds in the MOM_internal_tides code were modified to avoid
excessive calculations over the full data domain, some of which would use
uninitialized variables.  Also modified the logic determining the value of
diabatic_halo as returned by extract_diabatic_member to account for the wider
halos that are needed when the internal tide code is used.  This commit changes
the interfaces publicly visible routines and it changes answers in cases when
the internal_tides are used by doing calculations of the wave speeds in halo
points that are used in that code.