+(*)Non-Boussinesq default for Z_INIT_REMAP_GENERAL
  Changed the default value of Z_INIT_REMAP_GENERAL to true for fully
non-Boussinesq configurations.  All existing fully non-Boussinesq cases that use
INIT_LAYERS_FROM_Z_FILE = True use this setting, and it is likely that such
cases will not work at all if this is false.  This change reduces the parameters
that need to be changed to go between equivalent Boussinesq and non-Boussinesq
configurations to just BOUSSINESQ and SEMI_BOUSSINESQ.  The previous default
(false) is being retained for any Boussinesq or semi-Bousssinesq cases so that
all answers and output are bitwise identical in those cases, but by default some
non-Boussinesq solutions change answers, and there are changes to the
MOM_parameter_doc files for those non-Boussinesq cases.