makedep: Module dependency in nested includes
Nested includes are tracked for the purpose of include flags (-I), but
not with respect to the content within those files.  This patch tracks
the module usage statements (`use ...`) inside of any include files and
adds them to the Makefile rules of the top-level file.

This was implemented within the `nested_inc` function by adding a new
argument.