.testing: Codecov upload uses Github Actions token
Anonymous uploads to Codecov are throttled (though it is not clear if
this is happening on the Codecov or the GitHub Actions side).
Regardless, the advice to get around this throttling seems to be to use
the Codecov token associated with the project.

The .testing Makefile was modified to use this token if provided, and
the GitHub Actions environment now attempts to fetch this from the
secrets of the repository.

Hopefully individual groups can set this for their own projects, and it
will fall back to anonymous upload if unset, but I guess we'll have to
see how this plays out.