*Do 6 Newton method iterations in cuberoot
  Modified the cuberoot function to do 6 iterations with Newton's method,
starting from the optimal first guess of (3/8)**(1/3).  Following guidance from
performance testing of the previous version, all convergence testing has been
removed and the same number of iterations are applied regardless of the input
value.  The form of Newton's method used here does one division per iteration,
but it is good at polishing the root to give an accurate final solution and by
simply repeatedly using the same instructions it appears to optimize well.  This
commit changes answers at roundoff for code that uses the cuberoot function, so
ideally this PR would be dealt with before the cuberoot becomes widely used.