Testing: Remove mkmf macros for coupled tests
There appears to be a legacy macro and comments related to building the
coupled driver object files with mkmf.  To the best of my knowledge,
this is no longer used, and it appears to build correctly when removed,
so I am removing all reference to mkmf from this section.

However, I will pack it into a separate commit, in case it is actually
needed somewhere ^_^.