Better variables for updated masks in fill_miss_2d
  Within fill_miss_2d, there are 2 sets of masked values used for remapping -
masked values in neighboring points in arbitrary units (with names like "south")
and the updated valid mask with values of 0 or 1 (with names like "gs").  At one
point in the code, the former were being used where the latter were more
appropriate.  This commit changes this to help with the understandability of the
code.

  However, when the order of arithmetic is not specified with parentheses, the
Nvidia compiler is changing answers with this change of variables.  To avoid
this change in answers, this commit is only changing which variables are used
when the answer_date is larger than 20190101 and the order of the sums are
specified with parentheses.  This commit also now moves the logical test for
ans_2018 outside of the i- and j- loops where non-zero values of a_chg are set.

  All answers and output are bitwise identical.