+(*)Refactor Idealize_Hurricane
  Refactored the Idealized_Hurricane module to clean up strange or unscaled
variable units and eliminate dimensional scaling factors with the latest answer
dates.  This includes the introduction of 18 new runtime variables to replace
hard-coded dimensional constants and two runtime logical flags to reproduce
existing bugs.  An inconsistency in the sign convention for the distance from
the hurricane center with idealized_hurricane_wind_forcing in (the probably not
yet used) single column mode was also corrected. Also added descriptions with
units for all the variables in this module.  By default or with appropriate
parameter settings all answers are bitwise identical.