Use unscale arguments in 384 checksum calls
  Replaced 382 scale arguments with unscale arguments in calls to chksum
routines, and did the same in another 5 calls to global_mean or global_integral
routines.  Also added 2 missing unscale arguments to checksums for the initial
velocities, which probably had gone undetected because we usually initialize
velocities to 0.  All answers and diagnostic debugging output (apart from the
afore noted checksums of the initial velocities in some rare case) are bitwise
identical.