(*)Fix bug in rotate_ALE_sponge for fixed sponges
  Corrected a bug in rotate_ALE_sponge that incorrectly set the starting number
of sponge fields that prevented any cases with fixed ALE sponges from working
when ROTATE_INDEX is true, even if INDEX_TURNS is 0.  With this correction, the
code is now working (and giving identical answers) with rotation enabled for
cases with fixed ALE_sponges.  All answers are bitwise identical in cases that
worked before, but some cases that did not work at all are now working both
with and without grid rotation.