+Add and use query_debugging_checks
  Added the new function query_debugging_checks to the MOM_debugging module to
return flags indicating what types of debugging are enabled, including checksums
or redundant value checks.  This new routine is used to allow for the redundant
value checks in various step_MOM routines to be enabled or disabled at run-time
when debugging is on overall (by setting DEBUG = True).  For cases with
inconsistent redundant points, this change allows for a reduction in the volume
of debugging information sent to stdout from multiple PEs (in an arbitrary
order), while for cases without inconsistent redundant points this change will
allow for faster debugging with checksums alone.  All answers are bitwise
identical, but there is a new publicly visible routine.