Adds new pre-defined coordinate WOA09INT
In OM4, we use the "WOA09" coordinate for diagnostic z* output. This grid
was based on WOA09 but not exact because the WOA09 grid is not smooth.
We tried to construct a grid such that layer centers matched the depths
of WOA09 but this was not possible for six layers.

As an alternative, this new coordinate, "WOA09INT", uses the WOA09 depths
as interfaces so that the layer centers are always midway between the
WOA09 depths. To compare with WOA09 directly, WOA data now should be
interpolated but this is a simple half-half averaging.