Broke out testing into numerical_testing_type.F90
The helper type that shortens the testing lines will be needed
by both the upcoming Recon1d type and MOM_remapping, so it makes
sense to break the type out into a module. I've put this in framework/
but it is specifically a stand-alone module, not dependent on any
infrastructure code.