Streaming filter
The filters and their target frequencies are no longer hard-coded.
Instead, up to 10 filters with tidal frequencies as their target
frequencies and an unspecified number of filters with arbitrary
target frequencies can be turned on. The filter names are specified
in MOM_input and must consist of two letters/numbers. If a filter
name is the same as the name of a tidal constituent, then the
corresponding tidal frequency will be used as its target frequency.
Otherwise, the user must provide the target frequency. In either
case, the target frequency is specified by "TIDE_${FILTER_NAME}_FREQ"
in MOM_input.