Refactor spherical_harmonics_forward
  Refactored the spherical_harmonics_forward routine in MOM_spherical_harmonics
to work in rescaled units by making use of the unscale arguments to
reproducing_sum().  A total of 4 rescaling variables were moved into unscale
arguments, and a block of code that restores the scaling of the output variables
was eliminated.  The comments describing the units of several variables in this
module were made more explicit.  The two temporary arrays that store un-summed
contributions to the transforms were also moved out of the control structure and
made into local variables in the spherical_harmonics_forward routine to allow
for the reuse of that memory.  All answers and diagnostics are bitwise
identical, and no interfaces are changed.