MEKE: Conditionally compute biharmonic FrCoeff
This patch makes several adjustments to MOM_MEKE.F90 and
MOM_hor_visc.F90 to ensure that the Laplacian and biharmonic friction
coefficients are computed separately, and only if their respective terms
are enabled.

This resolves some subtle bugs where the default biharmonic value of -1
was applied to the Laplacian case, even when the biharmonic MEKE
friction was disabled.