+Add reset_h_var_halo_vals and reset_vector_halo_vals
  Added the new interfaces reset_h_var_halo_vals and reset_vector_halo_vals to
the MOM_debugging module to reset the halo regions of a 2-d or 3-d tracer-cell
array or the components of a 2-d or 3-d C-grid vector to specified values.  This
can be useful when working with arrays that have had their halo values replaced
with random values being used for intent(out) arrays, and for debugging bad
values along side checksum calls.  All answers are bitwise identical,
but there are two new public interfaces.