(*)Initialize halos for 7 intent out arrays
  Initialize the halos of 7 intent out arrays to 0 so that they will not have
uninitialized data in their halo regions.  The impacted arrays are u_cor and
v_cor in continuity_PPM, h_MLD in convert_MLD_to_ML_thickness, MLD in
energetic_PBL_get_MLD and LmixScale, bottomFac2 and barotrFac2 in
MEKE_lengthScales.  This does not change solutions in any of the MOM6-examples
test cases, but it is conceivable that could, and it does change the values
in the debugging checksum output for some cases to more nearly reproduce
under dimensional rescaling and rotation.